requirements.txt
flask==2.3.2
openai==1.51.2
python-dotenv==1.0.0
gunicorn==22.0.0